+++ src/commands/subsystem/txn/txn.ts
@@ -21,7 +21,7 @@ import { acquireDocsLock, DocsBusyError } from "../../../core/staging/lock.js";
 import { recoverInterruptedCommit } from "../../../core/staging/recovery/recover.js";
 import { commitStaged } from "../../../core/staging/txn/transaction.js";
 import { type PlannedWrite } from "../../../core/staging/writes.js";
-import { viewsAgree } from "../../../core/repo/tree/views.js";
+import { viewsAgree } from "../../../core/repo/tree/render/views.js";
 import { expectedViews } from "./views.js";
 
 export type ViewsAction = "created" | "updated" | "removed" | "current";